[INDY-2136] rename tests directory
